# docs/swagger.yaml
definitions:
...
  services.ServerErrorResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Service for handling credentials proofs (presentations)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Credential verification service API
  version: "1.0"
paths:
...
  /presentation/authorize:
    get:
      description: Handles the authorization request by checking the client_id and
        request_uri parameters, and then handling the request object
      parameters:
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Client ID
        in: query
        name: client_id
        required: true
        type: string
      - description: Request URI
        in: query
        name: request_uri
        required: true
        type: string
      - description: Auth URL
        in: query
        name: authUrl
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
      summary: Handles the authorization request
      tags:
      - external
...
  /presentation/proof/{id}/request-object/request.jwt:
    get:
      description: Responds with a request object by fetching the request object and
        setting it as fetched
      parameters:
      - description: Tenant ID
        in: path
        name: tenantId
        required: true
        type: string
      - description: Proof ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/jwt
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
      summary: Responds with a request object
      tags:
      - external
  /presentation/request:
    get:
      description: Handles the request for presentation by creating a presentation
        request with the provided parameters
      parameters:
      - description: Tenant ID
        in: query
        name: tenantId
        required: true
        type: string
      - description: Request ID
        in: query
        name: requestId
        required: true
        type: string
      - description: Group ID
        in: query
        name: groupId
        required: true
        type: string
      - description: TTL
        in: query
        name: ttl
        type: integer
      - description: Presentation Definition base64 url encoded
        in: query
        name: presentationDefinition
        required: true
        type: string
      produces:
      - application/jwt
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.ServerErrorResponse'
      summary: Handles the request for presentation
      tags:
      - external
swagger: "2.0"